+++ CAT/analysis/__init__.py
@@ -5,7 +5,6 @@ from .crs import (CRSJob, CRSResults)
 from .jobs import (job_single_point, job_geometry_opt, job_freq)
 from .ligand_bde import init_bde
 from .thermo_chem import (get_thermo, get_entropy)
-from .ligand_dissociate import dissociate_ligand
 from .ligand_solvation import init_solv
 
 
@@ -15,6 +14,5 @@ __all__ = [
     'job_single_point', 'job_geometry_opt', 'job_freq',
     'init_bde',
     'get_thermo', 'get_entropy',
-    'dissociate_ligand',
     'init_solv'
 ]