+++ CAT/utils.py
@@ -11,6 +11,23 @@ from os.path import join
 
 from scm.plams.core.settings import Settings
 
+from scm.plams.interfaces.adfsuite.ams import AMSJob
+from scm.plams.interfaces.adfsuite.adf import ADFJob
+from scm.plams.interfaces.thirdparty.orca import ORCAJob
+from scm.plams.interfaces.thirdparty.cp2k import Cp2kJob
+from scm.plams.interfaces.thirdparty.dirac import DiracJob
+from scm.plams.interfaces.thirdparty.gamess import GamessJob
+
+
+def type_to_string(job):
+    """ Turn a <type> object into a <str> object. """
+    job_dict = {ADFJob: 'adf', AMSJob: 'ams', DiracJob: 'dirac',
+                Cp2kJob: 'cp2k', GamessJob: 'gamess', ORCAJob: 'orca'}
+    try:
+        return job_dict[job]
+    except KeyError:
+        print(get_time() + 'WARNING: No default settings available for ' + str(job))
+
 
 def get_time():
     """ Returns the current time as string. """
@@ -46,12 +63,15 @@ def dict_concatenate(dic):
     return concact_dic
 
 
-def get_template(template_name):
+def get_template(template_name, from_cat_data=True):
     """
     Grab a template and return it as Settings object.
     """
-    path = join('data/templates', template_name)
-    xs = pkg.resource_string('CAT', path)
+    if from_cat_data:
+        path = join('data/templates', template_name)
+        xs = pkg.resource_string('CAT', path)
+    else:
+        xs = path
     if 'json' in template_name:
         s = json.loads(xs.decode())
     elif 'yaml' in template_name or 'yml' in template_name: