+++ CHANGELOG.rst
@@ -5,8 +5,14 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+0.3.2
+*****
+* Further (minor) updates and bug fixes to the database interaction
+* Overhaul of the bond dissociation energy (BDE) module
+* Job settings are now stored in the database
+
 0.3.0
-************
+*****
 
 * Massive overhaul of the CAT database interaction
 * Moved functions related to functiona group recognizition to CAT.attachment.ligand_anchoring.py